NEW_API: Add SingleDev communicator, Fix SYCL examples (#533)
* NEW_API: Add SingleDev communicator, Fix SYCL examples

+++ examples/include/sycl_base.hpp
@@ -7,10 +7,12 @@
 #include <iostream>
 #include <string>
 
+#include <mpi.h>
+
 #include <CL/sycl.hpp>
 #include "ccl.hpp"
 
-#define COUNT     (10 * 1024 * 1024)
+#define COUNT     size_t(10 * 1024 * 1024)
 #define COLL_ROOT (0)
 
 using namespace std;
